Refactor project structure, update styles, and hide API base URL in .env file for the React Client
