+++ client/src/App.css
@@ -15,12 +15,7 @@
 
 .App-header {
   background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
+  padding: 20px;
   color: white;
 }
 
@@ -37,6 +32,35 @@
   }
 }
 
+/* New styles for sticky footer */
+html, body, #root {
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
+
+.App-container {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+
+.App-main {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 20px;
+}
+
+.App-footer {
+  background-color: #0056b3;
+  color: white;
+  text-align: center;
+  padding: 10px 0;
+}
+
 /* Background Image */
 .background-image {
   position: fixed;
@@ -44,11 +68,10 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background-image: url('./assets/Disease_prediction.jpg'); /* Replace with your image path */
+  background-image: url('./assets/Disease_prediction.jpg');
   background-size: cover;
   background-position: center;
   filter: blur(8px);
   z-index: 1;
   opacity: 0.5;
 }
-
