+++ client/src/components/Footer.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+function Footer() {
+  return (
+    <footer
+      style={{
+        backgroundColor: '#0056b3',
+        color: 'white',
+        padding: '10px 0',
+        textAlign: 'center',
+        marginTop: '20px',
+      }}
+    >
+      &copy; 2024 Disease Prediction System. All rights reserved.
+    </footer>
+  );
+}
+
+export default Footer;
