+++ client/src/pages/Chatbot.js
@@ -20,7 +20,7 @@ function Chatbot() {
     setError(false);
 
     try {
-      const res = await fetch('http://localhost:5001/chatbot', {
+      const res = await fetch(`${process.env.REACT_APP_API_BASE_URL}/chatbot`, {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify({ query }),
